# docs/api/swagger.yaml
swagger: '2.0'
info:
  description: Entity service
  version: 1.0.0
  title: Well Delivery Service APIs
host: dns-host
basePath: /api/well-delivery
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/bhaRuns/v1/by_holeSection/{holeSection_id}':
    get:
      tags:
        - bhaRuns
      summary: Get a list of BHA Run objects for a hole section
      description: >-
        The API returns a list of BHA Run objects for a hole section id.  
        Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get a list of BHA Run objects for a hole section
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: holeSection_id
          in: path
          description: Hole section id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: BHA Run objects retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturns'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Hole section id or related BHA Run object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Not found BHA run with hole section id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/bhaRuns/v1/by_wellbore/{wellbore_id}:actual':
    get:
      tags:
        - bhaRuns
      summary: Get a list of actual BHA Run objects for a wellbore
      description: >-
        The API returns a list of actual BHA Run objects for a wellbore id.  
        Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get a list of actual BHA Run objects for a wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: wellbore_id
          in: path
          description: Wellbore id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Actual BHA Run objects retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturns'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Actual BHA Run object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Not found actual BHA run with wellbore id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
...
  '/drillingReports/v1/by_wellbore/{wellbore_id}':
    get:
      tags:
        - drillingReports
      summary: Get all available drilling report objects for the given wellbore id
      description: >-
        The API returns all available drilling report objects for the given
        wellbore id.   Required roles: service.storage.viewer,
        service.storage.creator or service.storage.admin.
      operationId: Get all available drilling report objects for the given wellbore id
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: wellbore_id
          in: path
          description: Wellbore id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Drilling report objects retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturns'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Drilling report object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Not found drilling reports with wellbore id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/drillingReports/v1/latest/by_wellbore/{wellbore_id}':
    get:
      tags:
        - drillingReports
      summary: Get latest drilling report object for the given wellbore id
      description: >-
        The API returns latest drilling report object for the given wellbore
        id.   Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get latest drilling report object forthe given wellbore id
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: wellbore_id
          in: path
          description: Wellbore id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Drilling report object retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturns'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Drilling report object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Not found drilling report with wellbore id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/drillingReports/v1/by_timeRange/{start_time}/{end_time}':
    get:
      tags:
        - drillingReports
      summary: Get a list of drilling report objects for the given time range
      description: >-
        The API returns a list of drilling report objects for the given time
        range. Date time format can be 'yyyyMMdd', 'yyyy-MM-dd', 'yyyy-M-d', 'yyyy-MM-ddTHH:mm:ss' or 'yyyy-MM-ddTHH:mm:ssZ'. Required roles: service.storage.viewer, service.storage.creator
        or service.storage.admin.
      operationId: Get a list of drilling report objects for the given time range
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: start_time
          in: path
          description: Start time
          required: true
          type: string
          default: '2021-03-01'
        - name: end_time
          in: path
          description: End time
          required: true
          type: string
          default: '2021-04-01'
      responses:
        '200':
          description: Drilling report objects retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturns'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Drilling report object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Not found drilling reports with wellbore id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/drillingReports/v1/reference_tree/by_drillingReport/{drilling_report_id}':
    get:
      tags:
        - drillingReports
      summary: Get the latest version of drilling report with a lists of URI references
      description: >-
        The API returns the latest version of well activity program with a lists
        of URI   references for a well. Required roles: service.storage.viewer,
        service.storage.creator or service.storage.admin.
      operationId: Get the latest version of drilling report with a lists of URI references
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: drilling_report_id
          in: path
          description: Drilling report id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Drilling report tree retrieved successfully.
          schema:
            $ref: '#/definitions/ReferenceTree'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Drilling report object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related drilling report reference tree with
                drilling report id: bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  AppError:
    type: object
    properties:
      message:
        type: string
...
  EntityReturn:
    type: object
    description: Entity return object
    properties:
      id:
        $ref: '#/definitions/EntityID'
      kind:
        $ref: '#/definitions/EntityKind'
      version:
        type: integer
        example: 1831253916104085
        format: int64
        description: The version number of this OSDU resource; set by the framework
      acl:
        $ref: '#/definitions/ACL'
      legal:
        $ref: '#/definitions/Legal'
      valid:
        type: boolean
        example: true
        description: The entity is valid or not.
      data:
        $ref: '#/definitions/EntityData'
  EntityReturns:
    type: array
    items:
      $ref: '#/definitions/EntityReturn'
  EntityID:
    type: string
    example: 'namespace:master-data--ExampleType:47c5967c-2961-55ea-96d7-e976a7954c26'
    description: The SRN which identifies this OSDU resource object without version.
    pattern: '^[\w\-\.]+:master-data\-\-ExampleType:[\w\-\.\:\%]+$'
  EntityKind:
    type: string
    example: 'slb-osdu-dev-des-prod-testing:well-delivery:exampletype:1.0.0'
    description: >-
      Kind of data it is being ingested and following the  pattern
      <Namespace>:<Source>:<Type>:<VersionMajor>.<VersionMinor>.<VersionPatch>. 
      The versioning scheme follows the semantic versioning,
      https://semver.org/..
    pattern: '^[\w\-\.]+:[\w\-\.]+:[\w\-\.]+:[0-9]+.[0-9]+.[0-9]+$'
  ACL:
    type: object
    description: The access control tags associated with this entity.
    required:
      - owners
      - viewers
    properties:
      viewers:
        type: array
        description: >-
          List of valid groups which will have view/read privileges over the
          entity.
        items:
          type: string
          example: data.default.viewers@slb-osdu-dev-des-prod-testing.contoso.com
      owners:
        type: array
        description: List of valid groups which will have write privileges over the entity.
        items:
          type: string
          example: data.default.owners@slb-osdu-dev-des-prod-testing.contoso.com
  Legal:
    type: object
    description: >-
      Attributes which represent the legal constraints associated with the
      record.
    properties:
      legaltags:
        type: array
        description: List of legaltag names associated with the entity.
        uniqueItems: true
        items:
          type: string
          example: sample-legal-tag
      otherRelevantDataCountries:
        type: array
        description: List of other relevant data countries.
        uniqueItems: true
        items:
          type: string
          example: US
          pattern: '^[A-Z]{2,2}$'
...
  EntityData:
    type: object
    description: Record payload represented as a list of key-value pairs.
    example:
      ExistenceKind: 'namespace:reference-data--ExistenceKind:planned:'
      NameWell: Well 01
      NameWellbore: Wellbore 01
  ReferenceTree:
    type: object
    properties:
      id:
        type: string
        example: ce49e5af-7f19-4625-a141-c0de4d5ac391
      version:
        type: integer
        example: 1599842774564
      type:
        type: string
        example: drillingprogram
      items:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            verion:
              type: integer
            type:
              type: string
            items:
              type: object
              properties:
                id:
                  type: string
                verion:
                  type: integer
                type:
                  type: string
        example:
          - id: bd932928-ff2c-4c8c-9f0d-b359325b77a7
            version: 1601570839996
            type: activityplan
          - id: bf61dbea-99ce-41ac-9498-6e9bc67f25c5
            version: 1601929333120
            type: rig
          - id: b3528018-860e-4766-a940-0adbcca8f63b
            version: 1601571043836
            type: well
          - id: 6a99fc9b-f9c0-4c8a-a100-a43b6b6d84ab
            version: 1601929331601
            type: wellbore
            items:
              - id: e947d79e-b721-4134-855f-e71c424097c8
                version: 1601927575971
                type: wellboreSegment
                items:
                  - id: 81336a4b-8547-4079-af07-ad3526b60585
                    version: 1601927575971
                    type: bharun
              - id: 8bfe58a2-7951-49bb-b1d1-e869b0b33db6
                version: 1601929345042
                type: wellboreSegment
                items:
                  - id: 2af558e2-d081-4296-913f-635a6b9226b9
                    version: 1601929346576
                    type: bharun
                  - id: 865bfed1-d197-4307-ba16-e94d2f86837f
                    version: 1601929347940
                    type: bharun